Removed the CDT.dll file
This Assembly provided encryption and registration key validation methods. The source for this assembly is not part of the OpenDental source, and the CDT.dll assembly provided with the source in fact does nothing.

A new project named 'OpenDental.Security' has been added to the project and will take over from CDT.dll. All methods in this library are placeholders. Anyone that wants to use Softlex Dental and use the encryption/registration facilities provided by the OpenDental.Security library will have to implement them themselves.