Refactored e-mail related code
Refactored up the e-mail related record classes, forms and database tables.

Support for direct messaging and webmail messages has been removed. In the base application I will only provide support for basic e-mail (e.g. via SMTP and POP3). Support for HTML mails has also been removed for now, it will be added back later once I find a better HTML editing component to use.

There are a number of forms of messaging that I would like to support. However rather than litter the base program with all the various forms of messaging I believe it will be better to aim to build a more generic 'framework' for messaging and enable plugins to tie in to this. This will be done in a future commit.

So in short:

- Refactored the e-mail related database tables.
- Refactored the e-mail related data classes and forms.
- Removed support for Direct Messaging and WebMail messaging.
- Marked a number of classes that are to be 'phased out' as obsolete (e.g. Lan, SIn, SOut, MsgBox, and a few others...)