Updated the way the translation helper looks up enum translations
The TranslationKeyAttribute has been renamed to TranslationAttribute and can be
used to define a default translation.

The helper method dynamically constructs a key in the format:
    E_<EnumTypeName>_<EnumStringValue>

For example:
    E_PreferenceName_AccountingSoftware
    E_DefinitionCategory_AccountColors

It will check if there is a value for that key in the language resource file.
If there is no translation for the key, it will check if the item has been
marked with a TranslationAttribute, and if so will use the translation string
from the attribute.